[bugfix] Add new condition for API of create user
In the lastest API of create user, I filter the special character.
The API of special character can only use Baseline(_) or Dot(.).
In order to let account could be accept email format,I add new condition
At(@) for the API.
And I modify the translated message,which is about create user with error special
character.